feat: widen quic_server_config_new max_early_data to u32
rustls 0.23.37 QUIC mode requires max_early_data_size ∈ {0, 0xFFFFFFFF};
the previous i32 signature with a negative-input guard could never
express the 0-RTT-enabled case. The FFI now rejects any other value at
the boundary, never propagating to rustls.

+++ navette/tls/config.mojo
@@ -177,7 +177,7 @@ struct QuicServerConfig(Movable):
         cert_pem: Span[UInt8, _],
         key_pem: Span[UInt8, _],
         alpn: String = "h3",
-        max_early_data: Int32 = 0,
+        max_early_data: UInt32 = UInt32(0),
     ) raises:
         self._lib = SharedLibrary(other=lib)
